+++ app/src/main/java/campus/tech/kakao/map/data/KakaoLocalApiService.kt
@@ -1,8 +1,7 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.data
 
 import retrofit2.Call
 import retrofit2.http.GET
-import retrofit2.http.Headers
 import retrofit2.http.Query
 
 interface KakaoLocalApiService {
