+++ app/src/main/res/layout/item_saved_keyword.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="keyword"
-            type="campus.tech.kakao.map.Keyword" />
+            type="campus.tech.kakao.map.data.Keyword" />
     </data>
 
     <LinearLayout
